Moved callbacks to their own submodule with prob_model | docs improvement
