+++ fortuna/prob_model/callbacks/base.py
