+++ fortuna/prob_model/fit_config/base.py
@@ -4,7 +4,7 @@ from fortuna.prob_model.fit_config.checkpointer import FitCheckpointer
 from fortuna.prob_model.fit_config.monitor import FitMonitor
 from fortuna.prob_model.fit_config.optimizer import FitOptimizer
 from fortuna.prob_model.fit_config.processor import FitProcessor
-from fortuna.prob_model.fit_config.callbacks import Callback
+from fortuna.prob_model.callbacks.base import Callback
 
 
 class FitConfig:
